gas-optimization
Storage Optimizations
Changed minOutPercentage and rewardDecayFactor to uint16 and packed them into a single storage slot
Updated corresponding events and function parameters for type consistency
Memory Caching
Added memory caching for storage variables in key functions:
_updateRewards, pendingRewards, _claimRewards, _deposit in Stake contract
_swapTokens in StakeSweeper contract
Unchecked Math
Added unchecked blocks for safe arithmetic operations to reduce gas costs
Applied to calculations that cannot overflow/underflow
Logic Improvements
Added early returns for common cases
Reduced storage reads/writes by batching operations
Improved variable naming and code organization
All tests pass, confirming optimizations maintain original functionality while improving efficiency.

+++ test/StakeSweeper.t.sol
@@ -18,7 +18,7 @@ contract StakeSweeperTest is Test {
     uint256 public constant INITIAL_SUPPLY = 1_000_000 ether;
 
     event RouterSet(address indexed router);
-    event MinOutPercentageSet(uint8 percentage);
+    event MinOutPercentageSet(uint16 percentage);
     event TokenSwept(address indexed token, uint256 amountIn, uint256 amountOut);
 
     function setUp() public {
@@ -81,7 +81,7 @@ contract StakeSweeperTest is Test {
     }
 
     function test_SetMinOutPercentage() public {
-        uint8 newPercentage = 95;
+        uint16 newPercentage = 95;
 
         vm.startPrank(owner);
